requirements.txt
streamlit
pandas
fpdf
pillow
st-gsheets-connection